refactor
para que todo tenga los nombres adecuados
