+++ Evaluador/.classpath
@@ -6,7 +6,6 @@
 		</attributes>
 	</classpathentry>
 	<classpathentry kind="src" path="src"/>
-	<classpathentry combineaccessrules="false" kind="src" path="/PracticasICI"/>
 	<classpathentry kind="lib" path="lib/ici.c1920.practica1.jar"/>
 	<classpathentry kind="lib" path="lib/ici.c2021.practica1.jar"/>
 	<classpathentry kind="lib" path="lib/ici-c2021-practica2.jar"/>
@@ -14,5 +13,8 @@
 	<classpathentry kind="lib" path="lib/PacManEngineFSMGraphObserver.jar"/>
 	<classpathentry kind="lib" path="lib/PacManEngineFuzzy.jar"/>
 	<classpathentry kind="lib" path="lib/jFuzzyLogic.jar"/>
+	<classpathentry combineaccessrules="false" kind="src" path="/PracticasICI"/>
+	<classpathentry kind="lib" path="lib/jess.jar"/>
+	<classpathentry kind="lib" path="lib/jsr94.jar"/>
 	<classpathentry kind="output" path="bin"/>
 </classpath>
