+++ Evaluador/src/config.properties
@@ -1,24 +1,7 @@
 trials = 1
 MsPacManTeams = \
-es.ucm.fdi.ici.practica4.demofuzzy.MsPacManFuzzy
+es.ucm.fdi.ici.c2021.practica4.grupo09.MsPacMan
 
 GhostsTeams = \
-es.ucm.fdi.ici.c2021.practica2.grupo01.Ghosts, \
-es.ucm.fdi.ici.c2021.practica2.grupo02.Ghosts, \
-es.ucm.fdi.ici.c2021.practica2.grupo03.Ghosts, \
-es.ucm.fdi.ici.c2021.practica2.grupo04.Ghosts, \
-es.ucm.fdi.ici.c2021.practica2.grupo05.Ghosts, \
-es.ucm.fdi.ici.c2021.practica2.grupo06.Ghosts, \
-es.ucm.fdi.ici.c2021.practica2.grupo07.Ghosts, \
-es.ucm.fdi.ici.c2021.practica2.grupo08.Ghosts, \
-es.ucm.fdi.ici.c2021.practica2.grupo09.Ghosts, \
-es.ucm.fdi.ici.c2021.practica1.grupo01.Ghosts, \
-es.ucm.fdi.ici.c2021.practica1.grupo02.Ghosts, \
-es.ucm.fdi.ici.c2021.practica1.grupo03.Ghosts, \
-es.ucm.fdi.ici.c2021.practica1.grupo04.Ghosts, \
-es.ucm.fdi.ici.c2021.practica1.grupo05.Ghosts, \
-es.ucm.fdi.ici.c2021.practica1.grupo06.Ghosts, \
-es.ucm.fdi.ici.c2021.practica1.grupo07.Ghosts, \
-es.ucm.fdi.ici.c2021.practica1.grupo08.Ghosts, \
-es.ucm.fdi.ici.c2021.practica1.grupo09.Ghosts, 
+es.ucm.fdi.ici.c2021.practica1.grupo01.Ghosts,
