+++ PracticasICI/src/main/java/es/ucm/fdi/ici/c2021/practica4/grupo09/MapaInfo.java
@@ -1,4 +1,4 @@
-package es.ucm.fdi.ici.practica4.demofuzzy;
+package es.ucm.fdi.ici.c2021.practica4.grupo09;
 import java.util.ArrayList;
 import java.util.EnumMap;
 import java.util.LinkedList;
