+++ PracticasICI/src/main/java/es/ucm/fdi/ici/c2021/practica4/grupo09/MsPacManActionSelector.java
@@ -1,12 +1,12 @@
-package es.ucm.fdi.ici.practica4.demofuzzy;
+package es.ucm.fdi.ici.c2021.practica4.grupo09;
 
 import java.util.HashMap;
 
+import es.ucm.fdi.ici.c2021.practica4.grupo09.actions.ChaseAction;
+import es.ucm.fdi.ici.c2021.practica4.grupo09.actions.ChillAction;
+import es.ucm.fdi.ici.c2021.practica4.grupo09.actions.RunAwayAction;
 import es.ucm.fdi.ici.fuzzy.Action;
 import es.ucm.fdi.ici.fuzzy.ActionSelector;
-import es.ucm.fdi.ici.practica4.demofuzzy.actions.ChaseAction;
-import es.ucm.fdi.ici.practica4.demofuzzy.actions.ChillAction;
-import es.ucm.fdi.ici.practica4.demofuzzy.actions.RunAwayAction;
 
 public class MsPacManActionSelector implements ActionSelector {
