+++ PracticasICI/src/main/java/es/ucm/fdi/ici/c2021/practica4/grupo09/MsPacManInput.java
@@ -1,4 +1,4 @@
-package es.ucm.fdi.ici.practica4.demofuzzy;
+package es.ucm.fdi.ici.c2021.practica4.grupo09;
 
 import java.util.HashMap;
