+++ PracticasICI/src/main/java/es/ucm/fdi/ici/c2021/practica4/grupo09/actions/EatPowePillAction.java
@@ -1,11 +1,11 @@
-package es.ucm.fdi.ici.practica4.demofuzzy.actions;
+package es.ucm.fdi.ici.c2021.practica4.grupo09.actions;
 
 
 import java.util.Random;
 
+import es.ucm.fdi.ici.c2021.practica4.grupo09.MapaInfo;
+import es.ucm.fdi.ici.c2021.practica4.grupo09.MapaInfo.interseccion;
 import es.ucm.fdi.ici.fuzzy.Action;
-import es.ucm.fdi.ici.practica4.demofuzzy.MapaInfo;
-import es.ucm.fdi.ici.practica4.demofuzzy.MapaInfo.interseccion;
 import pacman.game.Constants.DM;
 import pacman.game.Constants.GHOST;
 import pacman.game.Constants.MOVE;
