+++ PracticasICI/src/main/java/es/ucm/fdi/ici/c2021/practica4/grupo09/actions/RunAwayAction.java
@@ -1,8 +1,8 @@
-package es.ucm.fdi.ici.practica4.demofuzzy.actions;
+package es.ucm.fdi.ici.c2021.practica4.grupo09.actions;
 
 
+import es.ucm.fdi.ici.c2021.practica4.grupo09.MapaInfo;
 import es.ucm.fdi.ici.fuzzy.Action;
-import es.ucm.fdi.ici.practica4.demofuzzy.MapaInfo;
 import pacman.game.Constants.DM;
 import pacman.game.Constants.GHOST;
 import pacman.game.Constants.MOVE;
